chore: aggregate logic
