+++ internal/cmd/plugin/plugin_test.go
@@ -17,6 +17,13 @@ limitations under the License.
 package plugin
 
 import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8client "sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+
+	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
+	"github.com/cloudnative-pg/cloudnative-pg/internal/scheme"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
@@ -26,5 +33,47 @@ var _ = Describe("create client", func() {
 		err := createClient(cfg)
 		Expect(err).NotTo(HaveOccurred())
 		Expect(Client).NotTo(BeNil())
+	})
+})
+
+var _ = Describe("CompleteClusters", func() {
+	const namespace = "default"
+	var client k8client.Client
+
+	BeforeEach(func() {
+		cluster1 := &apiv1.Cluster{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "cluster1",
+				Namespace: namespace,
+			},
+		}
+		cluster2 := &apiv1.Cluster{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "cluster2",
+				Namespace: namespace,
+			},
+		}
+
+		client = fake.NewClientBuilder().WithScheme(scheme.BuildWithAllKnownScheme()).
+			WithObjects(cluster1, cluster2).Build()
+	})
+
+	It("should return matching cluster names", func(ctx SpecContext) {
+		toComplete := "clu"
+		result := completeClusters(ctx, client, namespace, toComplete)
+		Expect(result).To(HaveLen(2))
+		Expect(result).To(ConsistOf("cluster1", "cluster2"))
+	})
+
+	It("should return empty array when no clusters found", func(ctx SpecContext) {
+		toComplete := "nonexistent"
+		result := completeClusters(ctx, client, namespace, toComplete)
+		Expect(result).To(BeEmpty())
+	})
+
+	It("should skip clusters with prefix not matching toComplete", func(ctx SpecContext) {
+		toComplete := "nonexistent"
+		result := completeClusters(ctx, client, namespace, toComplete)
+		Expect(result).To(BeEmpty())
 	})
 })
